+++ docs/changelog.md
@@ -34,6 +34,7 @@ The main goals of v0.6 are reliability and performance. To achieve this required
  * `save-images` now supports image sequences
  * `-s`/`--stats` is now only used to write stats files, previously would load existing contents if any
     * Previous behavior can be obtained by merging stats files from multiple runs and setting the start/end times accordingly
+ * Using `--backend opencv` is now more robust to frame decode failures
 
 *v0.6-dev3*: